Time: 0 ms (100%), Space: 17.7 MB (20.52%) - LeetHub

+++ 0034-find-first-and-last-position-of-element-in-sorted-array/0034-find-first-and-last-position-of-element-in-sorted-array.cpp
@@ -8,7 +8,7 @@ public:
 
         while (start <= end) {
 
-            long long mid = start + (end - start) / 2;
+              int mid = start + (end - start) / 2;
 
             if (nums[mid] == target) {
                 ans = mid;
@@ -34,7 +34,7 @@ public:
 
         while(start <= end){
 
-             long long mid = start + (end - start)/2;
+              int mid = start + (end - start)/2;
 
              if(nums[mid] == target){
                 ans = mid;
